refactor + move rendering fns to own file
Merged redraw and redraw_ui into a single function, in order to make placeholder and password code easier

Moved most of redraw_common into main redraw function, having type wrangling issues with the image replace function
